+++ linuxcnc-hal/src/hal_pin/input_pin.rs
@@ -16,7 +16,7 @@ where
     ///
     /// Requires the full pin name including component like `vfd.speed-in` or `jog-pendant.enabled`.
     /// The component ID should be fetched from [`HalComponent.id`].
-    pub fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
+    pub(crate) fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
         let pin = P::register_pin(&name, PinDirection::In, component_id)?;
 
         Ok(Self { pin })