+++ linuxcnc-hal/src/hal_pin/output_pin.rs
@@ -17,7 +17,7 @@ where
     /// Requires the full pin name including component like `vfd.spindle-speed-out` or
     /// `jog-pendant.is-estopped`. The component ID should be fetched from
     /// [`HalComponentBuilder.id`].
-    pub fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
+    pub(crate) fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
         let pin = P::register_pin(&name, PinDirection::Out, component_id)?;
 
         Ok(Self { pin })