+++ linuxcnc-hal/src/hal_pin/bidirectional_pin.rs
@@ -1,35 +1,92 @@
-use crate::{
-    error::{PinRegisterError, StorageError},
-    hal_pin::{HalPin, PinDirection},
-};
+use crate::hal_pin::{pin_direction::PinDirection, PinRead, PinWrite};
+use linuxcnc_hal_sys::{hal_pin_bit_new, hal_pin_float_new, hal_pin_s32_new, hal_pin_u32_new};
 
-/// Wrapping struct to specialise a HAL pin to a bidirectional IO
-pub struct BidirectionalPin<P> {
-    pin: P,
+/// A pin that can be both read from and written to
+///
+/// Supported pin types are as follows
+///
+/// | Type                     | Storage | Equivalent `linuxcnc_hal_sys` function |
+/// | ------------------------ | ------- | -------------------------------------- |
+/// | `BidirectionalPin<f64>`  | `f64`   | [`hal_pin_float_new`]                  |
+/// | `BidirectionalPin<u32>`  | `u32`   | [`hal_pin_u32_new`]                    |
+/// | `BidirectionalPin<i32>`  | `i32`   | [`hal_pin_s32_new`]                    |
+/// | `BidirectionalPin<bool>` | `bool`  | [`hal_pin_bit_new`]                    |
+///
+/// # Examples
+///
+/// ## Create a pin
+///
+/// This example creates a `BidirectionalPin` under `demo-component.named-pin`.
+///
+/// ```rust,no_run
+/// use linuxcnc_hal::{
+///     hal_pin::BidirectionalPin,
+///     prelude::*,
+///     HalComponentBuilder,
+/// };
+/// use std::{thread, time::Duration, error::Error};
+///
+/// fn main() -> Result<(), Box<dyn Error>> {
+///     let mut builder = HalComponentBuilder::new("demo-component")?;
+///
+///     let pin = builder.register_pin::<BidirectionalPin<f64>>("named-pin")?;
+///
+///     let comp = builder.ready()?;
+///
+///     // Main control loop
+///     while !comp.should_exit() {
+///         println!("Input: {:?}", pin.value());
+///
+///         pin.set_value(123.45f64);
+///
+///         thread::sleep(Duration::from_millis(1000));
+///     }
+///
+///     Ok(())
+/// }
+/// ```
+pub struct BidirectionalPin<S> {
+    pub(crate) name: String,
+    pub(crate) storage: *mut *mut S,
 }
 
-impl<P> BidirectionalPin<P>
-where
-    P: HalPin,
-{
-    /// Register a new bidirectional pin with the HAL
-    ///
-    /// Requires the full pin name including component like `vfd.spindle-speed-out` or
-    /// `jog-pendant.is-estopped`. The component ID should be fetched from
-    /// [`HalComponentBuilder.id`].
-    pub fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
-        let pin = P::register_pin(&name, PinDirection::Bidirectional, component_id)?;
-
-        Ok(Self { pin })
-    }
-
-    /// Get the pin's value
-    pub fn value(&self) -> Result<&P::Storage, StorageError> {
-        self.pin.storage()
-    }
-
-    /// Set the pin's value
-    pub fn set_value(&self, value: P::Storage) -> Result<(), StorageError> {
-        Ok(*self.pin.storage_mut()? = value)
+impl<S> Drop for BidirectionalPin<S> {
+    fn drop(&mut self) {
+        debug!("Drop BidirectionalPin {}", self.name);
     }
 }
+
+impl_pin!(
+    BidirectionalPin,
+    f64,
+    hal_pin_float_new,
+    PinDirection::Bidirectional
+);
+impl_pin!(
+    BidirectionalPin,
+    u32,
+    hal_pin_u32_new,
+    PinDirection::Bidirectional
+);
+impl_pin!(
+    BidirectionalPin,
+    i32,
+    hal_pin_s32_new,
+    PinDirection::Bidirectional
+);
+impl_pin!(
+    BidirectionalPin,
+    bool,
+    hal_pin_bit_new,
+    PinDirection::Bidirectional
+);
+
+impl PinWrite for BidirectionalPin<f64> {}
+impl PinWrite for BidirectionalPin<u32> {}
+impl PinWrite for BidirectionalPin<i32> {}
+impl PinWrite for BidirectionalPin<bool> {}
+
+impl PinRead for BidirectionalPin<f64> {}
+impl PinRead for BidirectionalPin<u32> {}
+impl PinRead for BidirectionalPin<i32> {}
+impl PinRead for BidirectionalPin<bool> {}