+++ linuxcnc-hal/src/prelude.rs
@@ -1,0 +1,3 @@
+//! Prelude
+
+pub use crate::hal_pin::{PinRead, PinWrite};