+++ linuxcnc-hal/CHANGELOG.md
@@ -11,6 +11,25 @@ A safe abstraction used to create LinuxCNC HAL components in Rust.
 - #6 Added `BidirectionalPin` to allow an I/O pin to be registered on a component
 - Added basic `struct.rs` example showing storage of pins in a struct
 
+### Changed
+
+- **(breaking)** Simplify pin creation, reducing number of custom types. Pins are now created like this:
+
+  ```rust
+  use linuxcnc_hal::{
+      hal_pin::{InputPin, OutputPin},
+      prelude::*,
+      HalComponentBuilder,
+  };
+
+  let mut builder = HalComponentBuilder::new("pins")?;
+
+  let input_1 = builder.register_pin::<InputPin<f64>>("input-1")?;
+  let output_1 = builder.register_pin::<OutputPin<f64>>("output-1")?;
+  ```
+
+  Note the usage of `register_pin` as opposed to `register_input_pin` or `register_output_pin`.
+
 ## [0.1.3] - 2020-01-29
 
 ### Fixed