+++ linuxcnc-hal/CHANGELOG.md
@@ -6,6 +6,8 @@ A safe abstraction used to create LinuxCNC HAL components in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+## [0.2.0] - 2021-01-06
+
 ### Added
 
 - #6 Added `BidirectionalPin` to allow an I/O pin to be registered on a component
@@ -81,8 +83,9 @@ A safe abstraction used to create LinuxCNC HAL components in Rust.
 Initial release
 
 <!-- next-url -->
+[unreleased]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.2.0...HEAD
 
-[unreleased]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.1.3...HEAD
+[0.2.0]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.1.3...linuxcnc-hal-v0.2.0
 [0.1.3]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.1.2...linuxcnc-hal-v0.1.3
 [0.1.2]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.1.1...linuxcnc-hal-v0.1.2
 [0.1.1]: https://github.com/jamwaffles/linuxcnc-hal-rs/compare/linuxcnc-hal-v0.1.0...linuxcnc-hal-v0.1.1